Refactor server
Tests were passing but also throwing errors at the end. Turned out that binding the server to a specific port is
keeping the server open, and unable to close after tests are completed.
The solution was to split the server into separate files,bind only production server to a specific port and
allow supertest to choose it's own port for the testing.